+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 
 #### Version: 0.2.0-SNAPSHOT
-- compose transducers with `+`
-- transduce accepts sequences as input
-- `cat()` transducer signature more explicit (vs type inference)
+- compose transducers with `+`.
+- transduce accepts sequences as input.
+- `cat()` transducer signature more explicit (vs type inference).
+- **Fix**: `cat()` invoked completion operation after each input value.  
 
 #### Version: 0.1.0 - 2017-01-19
 - initial release